+++ History.md
@@ -1,3 +1,10 @@
+2.0.0 / 2014-10-08
+==================
+
+ * Insert version into component id
+ * Update footprint of `.compile` and `.compileAll`
+
+
 1.0.1 / 2014-09-15
 ==================